+++ src/components/DpMultiselect/DpMultiselect.vue
@@ -83,6 +83,7 @@
 </template>
 
 <script>
+import { dpValidateMultiselectDirective } from '../../lib/validation'
 import VueMultiselect from 'vue-multiselect'
 
 export default {
@@ -90,6 +91,10 @@ export default {
 
   components: {
     VueMultiselect
+  },
+
+  directives: {
+    dpValidateMultiselectDirective
   },
 
   props: {